Task3_Weather_App/requirements.txt
requests==2.32.0
Pillow==10.0.1
